// backend/src/constants/enum.ts
export enum Games {
    BINGO = "bingo"
}

export enum SocketChannels {
    TURN = "turn",
    START_GAME = "start-game",
    PLAYER_JOINED = "player-joined",
    WINNER = "winner",
    STRIKE_UPDATE = "strike-update",
    CHAT_MESSAGE = "chat-message",
    CHAT_HISTORY = "chat-history",
    REPLAY_ROUND = "replay-round",
    LAST_PLAYER_STANDING = "last-player-standing"
}